feat: 컴포넌트 재사용을 위한 분리 및 세부 간격 조정
